+++ internal/cri/util/references.go
@@ -16,7 +16,10 @@
 
 package util
 
-import reference "github.com/distribution/reference"
+import (
+	reference "github.com/distribution/reference"
+	imagedigest "github.com/opencontainers/go-digest"
+)
 
 // ParseImageReferences parses a list of arbitrary image references and returns
 // the repotags and repodigests
@@ -34,4 +37,19 @@ func ParseImageReferences(refs []string) ([]string, []string) {
 		}
 	}
 	return tags, digests
+}
+
+// GetRepoDigestAndTag returns image repoDigest and repoTag of the named image reference.
+func GetRepoDigestAndTag(namedRef reference.Named, digest imagedigest.Digest, schema1 bool) (string, string) {
+	var repoTag, repoDigest string
+	if _, ok := namedRef.(reference.NamedTagged); ok {
+		repoTag = namedRef.String()
+	}
+	if _, ok := namedRef.(reference.Canonical); ok {
+		repoDigest = namedRef.String()
+	} else if !schema1 {
+		// digest is not actual repo digest for schema1 image.
+		repoDigest = namedRef.Name() + "@" + digest.String()
+	}
+	return repoDigest, repoTag
 }
